ref(debuginfo): Move symbol table logic from Object to own trait
