+++ debuginfo/src/lib.rs
@@ -1,14 +1,18 @@
 //! Abstraction for reading debug info files.
 
-extern crate uuid;
 extern crate goblin;
-#[macro_use] extern crate symbolic_common;
-#[macro_use] extern crate if_chain;
+#[macro_use]
+extern crate if_chain;
+#[macro_use]
+extern crate symbolic_common;
+extern crate uuid;
 
 mod breakpad;
-mod object;
 mod dwarf;
+mod object;
+mod symbols;
 
 pub use breakpad::*;
 pub use object::*;
 pub use dwarf::*;
+pub use symbols::*;
